Added Command Line menu + search properties to buy
I added a command menu which the user can enter number 1, 2 or 3, 1 being search for rental properties, 2 being search for properties to buy, and 3 being exit the script. 

Once option 1 or 2 are selected, the user will be prompted to enter the area they which to search for properties to buy or rent, e,g. Newtownabbey.

Scraping will commence. Once scraping is complete, the command line will prompt user to enter filename they which to save, e.g. newtownabbeyrentals

Once this is entered an .csv file will be created with the filename entered by the user, in the same directory as this script. This file will contain the scraped data.