Fixed Issue wth rental scrapes appending to sale scrapers file
There seemed to be an issue when you would scrape rental properties, and afterwards scrape properties to buy, within the properties to buy outputted .csv file the rental properties would be present in this file as well. This of course was not correct. This has been rectified by making two separate lists for houses to rent and houses for sale.