Took out some repeated code and add functions, fixed error with previously scraped data being appended to outputted file
I removed repeated code, such as the for g_data loop, which I had two identical pieces of code for the rental and buy parts of the scraper. I have not put this into a function to hopefully make code more readable, and to prevent repeated code.

I also noticed an error in that if the user was to select Rent Search (1) for example, enter the town and save the .csv file, once the menu appeared again and the user selected Rent Search and entered another town, the outputted .csv file for this second scraped town would also contain the scraped data from the previous search. I think the issue was as the first set of data was appended to the list, when the second set of data of scraped and appended to the list the list still contained the first set of scraped data. To solve this I made a function which I run after each .csv is saved to delete the data contained within the list.