Fixed error when scraping properties to buy
Error was that as properties for sale do not have period of payment, e.g. per month, the scraper was having a 'period' not defined error, to correct this I had to add a new function for the properties to buy that didn't scrape the period section of the site, so the error would not occur.